Update the persistent timestamp even if there are no oplog entries
We use the oplog timestamp recorded in the db as an indicator of
whether our oplog tailers are up to date or not. However, on the rare
occasion when we have a totally idle database, the timestamps don't
get updated.

This change ensures that the timestamp is updated, even when new
entries aren't being written to the oplog. As long as we grab the
timestamp before querying the oplog, if we then query the oplog and
get no results, we know there won't be any entries older than the
current server timestamp.